+++ registry/routes/component-preview.js
@@ -6,19 +6,18 @@ var urlBuilder = require('../domain/url-builder');
 
 module.exports = function(repository){
   return function(req, res){
-
+    
     repository.getComponent(req.params.componentName, req.params.componentVersion, function(err, component){
 
       if(err){
         res.errorDetails = err;
+        res.errorCode = 'NOT_FOUND';
         return res.json(404, { err: err });
       }
 
       var isHtmlRequest = !!req.headers.accept && req.headers.accept.indexOf('text/html') >= 0;
 
       if(isHtmlRequest && !!res.conf.discovery){
-
-        
                 
         return res.render('component-preview', {
           component: component,
